requirements.txt
streamlit
opencv-python-headless
pillow
mediapipe
ultralytics
numpy
torch